backend/requirements.txt
# ================================
# Web framework
# ================================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ================================
# Base de datos
# ================================
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pgvector==0.3.5

# ================================
# Autenticación
# ================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]==2.9.2
pydantic-settings==2.6.0

# ================================
# Cola de tareas async
# ================================
celery==5.4.0
redis==5.1.1

# ================================
# Storage (S3 / MinIO)
# ================================
boto3==1.35.44

# ================================
# Procesamiento de documentos
# ================================
pymupdf==1.24.12
python-docx==1.1.2

# ================================
# AI / RAG
# ================================
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
langgraph==0.2.45
openai>=1.54.0,<2.0.0
tiktoken==0.8.0

# ================================
# ML
# ================================
scikit-learn==1.5.2
numpy==1.26.4
sentence-transformers==3.2.1

# ================================
# Testing
# ================================
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# ================================
# Utils
# ================================
python-dotenv==1.0.1
structlog==24.4.0